perf: 修改 account config
