+++ apps/accounts/models/__init__.py
@@ -1,3 +1,4 @@
 from .account import *
 from .automations import *
 from .base import *
+from .template import *
