+++ apps/common/serializers/dynamic.py
@@ -51,8 +51,8 @@ def create_serializer_class(serializer_name, fields_info):
         field_type = data.pop('type', 'str')
 
         # 用户定义 default 和 required 可能会冲突, 所以要处理一下
-        default = data.get('default', '')
-        if default not in ['', None]:
+        default = data.get('default', (None, None))
+        if default is not (None, None):
             data['required'] = False
         else:
             data.pop('default', None)
