+++ apps/locale/ja/LC_MESSAGES/django.mo
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:6d560f3d093c8e632332ff432379d03cc1ebc7b191e1393625dafe9d15c1dda2
-size 146083
+oid sha256:d48927f8cc38ff8c2ac224196206a7464b7603ac4916c7ab4391d3e64165d92f
+size 147265
