+++ apps/locale/zh/LC_MESSAGES/django.mo
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:2f7feedee9439ae2d31445855ec49053f1f6c87922c5452ac6a312c8633605be
-size 119333
+oid sha256:1583415690d2a6e91b350a31e68203fc733f0e5fed6a8f8e45d3d134f1bd5e23
+size 120268
